requirements.txt
python-docx
pandas
python-dotenv
openai
python-telegram-bot
reportlab
Pillow
python-pptx
openpyxl